% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{AcknowledgeRequest}
\alias{AcknowledgeRequest}
\title{AcknowledgeRequest Object}
\usage{
AcknowledgeRequest(ackIds = NULL)
}
\arguments{
\item{ackIds}{The acknowledgment ID for the messages being acknowledged that was returned}
}
\value{
AcknowledgeRequest object
}
\description{
AcknowledgeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the Acknowledge method.
}
\seealso{
Other AcknowledgeRequest functions: \code{\link{projects.subscriptions.acknowledge}}
}
