% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListTopicsResponse}
\alias{ListTopicsResponse}
\title{ListTopicsResponse Object}
\usage{
ListTopicsResponse(topics = NULL, nextPageToken = NULL)
}
\arguments{
\item{topics}{The resulting topics}

\item{nextPageToken}{If not empty, indicates that there may be more topics that match the request; this value should be passed in a new `ListTopicsRequest`}
}
\value{
ListTopicsResponse object
}
\description{
ListTopicsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListTopics` method.
}

