% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_functions.R
\name{projects.topics.getIamPolicy}
\alias{projects.topics.getIamPolicy}
\title{Gets the access control policy for a `resource`. Returns an empty policy if the resource exists and does not have a policy set.}
\usage{
projects.topics.getIamPolicy(resource)
}
\arguments{
\item{resource}{REQUIRED: The resource for which the policy is being requested}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/pubsub
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/pubsub)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/pubsub/docs}{Google Documentation}
}

