% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PublishRequest}
\alias{PublishRequest}
\title{Google Cloud Pub/Sub API Objects 
Provides reliable, many-to-many, asynchronous messaging between applications.}
\usage{
PublishRequest(messages = NULL)
}
\arguments{
\item{messages}{The messages to publish}
}
\value{
PublishRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:07:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlepubsubv1.auto/R/pubsub_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
PublishRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the Publish method.
}
\seealso{
Other PublishRequest functions: \code{\link{projects.topics.publish}}
}
