% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PublishResponse}
\alias{PublishResponse}
\title{PublishResponse Object}
\usage{
PublishResponse(messageIds = NULL)
}
\arguments{
\item{messageIds}{The server-assigned ID of each published message, in the same order as}
}
\value{
PublishResponse object
}
\description{
PublishResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `Publish` method.
}
