% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PullResponse}
\alias{PullResponse}
\title{PullResponse Object}
\usage{
PullResponse(receivedMessages = NULL)
}
\arguments{
\item{receivedMessages}{Received Pub/Sub messages}
}
\value{
PullResponse object
}
\description{
PullResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `Pull` method.
}

