% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{PushConfig}
\alias{PushConfig}
\title{PushConfig Object}
\usage{
PushConfig(PushConfig.attributes = NULL, pushEndpoint = NULL,
  attributes = NULL)
}
\arguments{
\item{PushConfig.attributes}{The \link{PushConfig.attributes} object or list of objects}

\item{pushEndpoint}{A URL locating the endpoint to which messages should be pushed}

\item{attributes}{Endpoint configuration attributes}
}
\value{
PushConfig object
}
\description{
PushConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for a push delivery endpoint.
}
\seealso{
Other PushConfig functions: \code{\link{PushConfig.attributes}}
}
