% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{SetIamPolicyRequest}
\alias{SetIamPolicyRequest}
\title{SetIamPolicyRequest Object}
\usage{
SetIamPolicyRequest(policy = NULL)
}
\arguments{
\item{policy}{REQUIRED: The complete policy to be applied to the `resource`}
}
\value{
SetIamPolicyRequest object
}
\description{
SetIamPolicyRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `SetIamPolicy` method.
}
\seealso{
Other SetIamPolicyRequest functions: \code{\link{projects.snapshots.setIamPolicy}},
  \code{\link{projects.subscriptions.setIamPolicy}},
  \code{\link{projects.topics.setIamPolicy}}
}
