% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{Topic}
\alias{Topic}
\title{Topic Object}
\usage{
Topic(name = NULL)
}
\arguments{
\item{name}{The name of the topic}
}
\value{
Topic object
}
\description{
Topic Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A topic resource.
}
\seealso{
Other Topic functions: \code{\link{projects.topics.create}}
}

