% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ListTopicSubscriptionsResponse}
\alias{ListTopicSubscriptionsResponse}
\title{ListTopicSubscriptionsResponse Object}
\usage{
ListTopicSubscriptionsResponse(nextPageToken = NULL, subscriptions = NULL)
}
\arguments{
\item{nextPageToken}{If not empty, indicates that there may be more subscriptions that match}

\item{subscriptions}{The names of the subscriptions that match the request}
}
\value{
ListTopicSubscriptionsResponse object
}
\description{
ListTopicSubscriptionsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListTopicSubscriptions` method.
}
