% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{ModifyAckDeadlineRequest}
\alias{ModifyAckDeadlineRequest}
\title{ModifyAckDeadlineRequest Object}
\usage{
ModifyAckDeadlineRequest(ackDeadlineSeconds = NULL, ackIds = NULL)
}
\arguments{
\item{ackDeadlineSeconds}{The new ack deadline with respect to the time this request was sent to}

\item{ackIds}{List of acknowledgment IDs}
}
\value{
ModifyAckDeadlineRequest object
}
\description{
ModifyAckDeadlineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the ModifyAckDeadline method.
}
\seealso{
Other ModifyAckDeadlineRequest functions: \code{\link{projects.subscriptions.modifyAckDeadline}}
}
