% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_objects.R
\name{SetIamPolicyRequest}
\alias{SetIamPolicyRequest}
\title{Google Cloud Pub/Sub API Objects 
Provides reliable, many-to-many, asynchronous messaging between applications.}
\usage{
SetIamPolicyRequest(policy = NULL)
}
\arguments{
\item{policy}{REQUIRED: The complete policy to be applied to the `resource`}
}
\value{
SetIamPolicyRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:22:16
filename: /Users/mark/dev/R/autoGoogleAPI/googlepubsubv1.auto/R/pubsub_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SetIamPolicyRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `SetIamPolicy` method.
}
\seealso{
Other SetIamPolicyRequest functions: \code{\link{projects.subscriptions.setIamPolicy}},
  \code{\link{projects.topics.setIamPolicy}}
}

